:white_check_mark: #29 Move to the warning.spec
